+++ app/components/UI/Notification/index.js
@@ -1,50 +1,18 @@
 import React, { useEffect, useState, useRef, useMemo } from 'react';
-import { StyleSheet } from 'react-native';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import Animated, { Easing } from 'react-native-reanimated';
 import { hideTransactionNotification } from '../../../actions/notification';
-import { colors } from '../../../styles/common';
 import notificationTypes from '../../../util/notifications';
-import BaseNotification from './BaseNotification';
-import Device from '../../../util/Device';
-import ElevatedView from 'react-native-elevated-view';
 import TransactionNotification from './TransactionNotification';
+import SimpleNotification from './SimpleNotification';
 
 const { TRANSACTION, SIMPLE } = notificationTypes;
 
 const BROWSER_ROUTE = 'BrowserView';
 
-const styles = StyleSheet.create({
-	modalTypeView: {
-		position: 'absolute',
-		bottom: 0,
-		paddingBottom: Device.isIphoneX() ? 20 : 10,
-		left: 0,
-		right: 0,
-		backgroundColor: colors.transparent
-	},
-	modalTypeViewBrowser: {
-		bottom: Device.isIphoneX() ? 70 : 60
-	},
-	notificationContainer: {
-		flex: 0.1,
-		flexDirection: 'row',
-		alignItems: 'flex-end'
-	}
-});
-
 function Notification(props) {
-	const {
-		autodismiss,
-		isVisible,
-		navigation,
-		hideTransactionNotification,
-		notificationTitle,
-		notificationDescription,
-		notificationStatus,
-		notificationType
-	} = props;
+	const { autodismiss, isVisible, navigation, hideTransactionNotification, notificationType } = props;
 
 	const [internalIsVisible, setInternalIsVisible] = useState(isVisible);
 
@@ -76,20 +44,6 @@ function Notification(props) {
 		return route.routeName === BROWSER_ROUTE;
 	}, [navigation.state]);
 
-	const handleSimpleNotification = () => (
-		<ElevatedView style={[styles.modalTypeView, isInBrowserView && styles.modalTypeViewBrowser]} elevation={100}>
-			<Animated.View
-				style={[styles.notificationContainer, { transform: [{ translateY: notificationAnimated }] }]}
-			>
-				<BaseNotification
-					status={notificationStatus}
-					data={{ title: notificationTitle, description: notificationDescription }}
-					onHide={hideTransactionNotification}
-				/>
-			</Animated.View>
-		</ElevatedView>
-	);
-
 	useEffect(() => {
 		hideTransactionNotification();
 	}, [hideTransactionNotification]);
@@ -97,17 +51,12 @@ function Notification(props) {
 	useEffect(() => {
 		if (!prevIsVisible && isVisible) {
 			// Auto dismiss notification in case of confirmations
-			autodismiss &&
-				setTimeout(() => {
-					hideTransactionNotification();
-				}, autodismiss);
+			autodismiss && setTimeout(() => hideTransactionNotification(), autodismiss);
 			setInternalIsVisible(true);
 			setTimeout(() => animatedTimingStart(notificationAnimated, 0), 100);
 		} else if (prevIsVisible && !isVisible) {
 			animatedTimingStart(notificationAnimated, 200);
-			setTimeout(() => {
-				setInternalIsVisible(false);
-			}, 500);
+			setTimeout(() => setInternalIsVisible(false), 500);
 		}
 	}, [
 		isVisible,
@@ -129,7 +78,8 @@ function Notification(props) {
 				animatedTimingStart={animatedTimingStart}
 			/>
 		);
-	if (notificationType === SIMPLE) return handleSimpleNotification();
+	if (notificationType === SIMPLE)
+		return <SimpleNotification isInBrowserView={isInBrowserView} notificationAnimated={notificationAnimated} />;
 	return null;
 }
 
@@ -151,18 +101,6 @@ Notification.propTypes = {
 	 */
 	hideTransactionNotification: PropTypes.func,
 	/**
-	 * Title for notification if defined
-	 */
-	notificationTitle: PropTypes.string,
-	/**
-	 * Description for notification if defined
-	 */
-	notificationDescription: PropTypes.string,
-	/**
-	 * Status for notification if defined
-	 */
-	notificationStatus: PropTypes.string,
-	/**
 	 * Type of notification, transaction or simple
 	 */
 	notificationType: PropTypes.string
@@ -172,9 +110,6 @@ const mapStateToProps = state => ({
 	accounts: state.engine.backgroundState.AccountTrackerController.accounts,
 	isVisible: state.notification.isVisible,
 	autodismiss: state.notification.autodismiss,
-	notificationTitle: state.notification.title,
-	notificationStatus: state.notification.status,
-	notificationDescription: state.notification.description,
 	notificationType: state.notification.type
 });
